Change config floats to doubles

+++ AutoStationConfig/Config.cs
@@ -15,7 +15,7 @@ namespace Pasukaru.DSP.AutoStationConfig
 
             public static ConfigEntry<int> DroneTransportRange;
             public static ConfigEntry<int> MinDroneLoad;
-            public static ConfigEntry<float> DroneInsertPercentage;
+            public static ConfigEntry<double> DroneInsertPercentage;
         }
 
         public static class ILS
@@ -24,11 +24,11 @@ namespace Pasukaru.DSP.AutoStationConfig
 
             public static ConfigEntry<int> DroneTransportRange;
             public static ConfigEntry<int> MinDroneLoad;
-            public static ConfigEntry<float> DroneInsertPercentage;
+            public static ConfigEntry<double> DroneInsertPercentage;
 
             public static ConfigEntry<int> VesselTransportRange;
             public static ConfigEntry<int> MinVesselLoad;
-            public static ConfigEntry<float> VesselInsertPercentage;
+            public static ConfigEntry<double> VesselInsertPercentage;
 
             public static ConfigEntry<double> MinWarpDistance;
             public static ConfigEntry<bool> WarperInLastItemSlot;
@@ -77,10 +77,10 @@ namespace Pasukaru.DSP.AutoStationConfig
                 )
             );
 
-            PLS.DroneInsertPercentage = config.Bind(PLS_SECTION, "Drone Insert Percentage", 1f,
+            PLS.DroneInsertPercentage = config.Bind(PLS_SECTION, "Drone Insert Percentage", 1d,
                 new ConfigDescription(
                     "Amount of drones to insert. For vanilla ILS, 0.05 = 5% = 50/100*0.5 = 2.5, rounded to 3.",
-                    new AcceptableValueRange<float>(0, 1),
+                    new AcceptableValueRange<double>(0, 1),
                     new { }
                 )
             );
@@ -113,10 +113,10 @@ namespace Pasukaru.DSP.AutoStationConfig
                 )
             );
 
-            ILS.DroneInsertPercentage = config.Bind(ILS_SECTION, "Drone Insert Percentage", 1f,
+            ILS.DroneInsertPercentage = config.Bind(ILS_SECTION, "Drone Insert Percentage", 1d,
                 new ConfigDescription(
                     "Amount of drones to insert. For vanilla ILS, 0.05 = 5% = 50/100*0.5 = 2.5, rounded to 3.",
-                    new AcceptableValueRange<float>(0, 1),
+                    new AcceptableValueRange<double>(0, 1),
                     new { }
                 )
             );
@@ -138,10 +138,10 @@ namespace Pasukaru.DSP.AutoStationConfig
                 )
             );
             
-            ILS.VesselInsertPercentage = config.Bind(ILS_SECTION, "Vessel Insert Percentage", 1f,
+            ILS.VesselInsertPercentage = config.Bind(ILS_SECTION, "Vessel Insert Percentage", 1d,
                 new ConfigDescription(
                     "Amount of vessels to insert. 0.01 = 1%. For vanilla ILS, 0.05 => 10/100*5 = 0.5, rounded to 1.",
-                    new AcceptableValueRange<float>(0, 1),
+                    new AcceptableValueRange<double>(0, 1),
                     new { }
                 )
             );
